requirements.txt
torch
torchvision
opencv-python
pyyaml
json_tricks
yacs
scikit-learn
pandas
timm==0.4.12
numpy==1.23.5
einops
fvcore
transformers
#==4.19.2
sentencepiece
ftfy
regex
nltk
vision-datasets==0.2.2
pycocotools
diffdist
pyarrow
cityscapesscripts
shapely
scikit-image
mup
#gradio==3.35.2
scann
kornia==0.6.4
torchmetrics==0.6.0
progressbar
pillow==9.4.0
